move _make_path_function, _check_path_function to utility
