+++ signac/__main__.py
@@ -35,8 +35,8 @@ from .common import config
 from .common.configobj import Section, flatten_errors
 from .common.crypt import get_crypt_context, get_keyring, parse_pwhash
 from .contrib.filterparse import parse_filter_arg
-from .contrib.import_export import _SchemaPathEvaluationError, export_jobs
-from .contrib.utility import add_verbosity_argument, prompt_password, query_yes_no
+from .contrib.import_export import export_jobs
+from .contrib.utility import add_verbosity_argument, prompt_password, query_yes_no, _SchemaPathEvaluationError
 from .diff import diff_jobs
 from .errors import (
     DestinationExistsError,
