+++ signac/contrib/import_export.py
@@ -11,14 +11,13 @@ import re
 import shutil
 import tarfile
 import zipfile
-from collections import OrderedDict
 from contextlib import closing, contextmanager
 from string import Formatter
 from tempfile import TemporaryDirectory
 from zipfile import ZIP_DEFLATED, ZipFile
 
 from .errors import DestinationExistsError, StatepointParsingError
-from .utility import _dotted_dict_to_nested_dicts, _mkdir_p
+from .utility import _check_path_function, _make_path_function, _dotted_dict_to_nested_dicts, _mkdir_p
 
 logger = logging.getLogger(__name__)
 
@@ -35,258 +34,6 @@ RE_TYPES = {
 
 
 #  ### Export related  ###
-
-
-def _make_schema_based_path_function(jobs, exclude_keys=None, delimiter_nested="."):
-    """Generate schema-based paths as a function of the given jobs.
-
-    Parameters
-    ----------
-    jobs : iterable of :class:`~signac.contrib.job.Job`
-        A sequence of jobs (instances of :class:`~signac.contrib.job.Job`).
-    exclude_keys : sequence[str]
-        A sequence of keys to exclude (Default value = None).
-    delimiter_nested : str
-        Delimiter used for nesting keys (Default value = '.').
-
-    Returns
-    -------
-    callable
-        Function that returns a normalized path.
-
-    """
-    from .schema import _build_job_statepoint_index
-
-    if len(jobs) <= 1:
-        # The lambda must (optionally) take a format spec argument to match the
-        # signature of the path function below.
-        return lambda job, sep=None: ""
-
-    index = [{"_id": job.id, "sp": job.sp()} for job in jobs]
-    jsi = _build_job_statepoint_index(exclude_const=True, index=index)
-    sp_index = OrderedDict(jsi)
-
-    paths = {}
-    for key_tokens, values in sp_index.items():
-        key = key_tokens.replace(".", delimiter_nested)
-        if exclude_keys and key in exclude_keys:
-            continue
-        for value, group in values.items():
-            path_tokens = key, str(value)
-            for job_id in group:
-                paths.setdefault(job_id, list())
-                paths[job_id].extend(path_tokens)
-
-    def path(job, sep=None):
-        """Normalize the path.
-
-        Parameters
-        ----------
-        job : :class:`~signac.contrib.job.Job`
-            An instance of :class:`~signac.contrib.job.Job`.
-        sep : str
-            (Default value = None)
-
-        Returns
-        -------
-        str
-            Normalized path.
-
-        Raises
-        ------
-        RuntimeError
-            If unable to determine path for job.
-
-        """
-        try:
-            if sep:
-                return os.path.normpath(sep.join(paths[job.id]))
-            else:
-                return os.path.normpath(os.path.join(*paths[job.id]))
-        except KeyError:
-            raise RuntimeError(
-                "Unable to determine path for job '{}'.\nThis is usually caused by a "
-                "heterogeneous schema, where some keys are only present in some jobs. "
-                "Try providing a custom path.".format(job)
-            )
-
-    return path
-
-
-class _AutoPathFormatter(Formatter):
-    """A custom formatter used to format jobs with their respective paths."""
-
-    def __init__(self, paths):
-        self.paths = paths
-
-    def format_field(self, value, format_spec):
-        """Format string fields, replacing jobs with their paths.
-
-        Parameters
-        ----------
-        value :
-            The value to be formatted.
-        format_spec :
-            The format specification.
-
-        Returns
-        -------
-        str
-            Formatted string.
-
-        """
-        from .job import Job
-
-        if isinstance(value, Job):
-            return self.paths(value, format_spec)
-        else:
-            return super().format_field(value, format_spec)
-
-
-class _SchemaPathEvaluationError(RuntimeError):
-    """Raised for errors in schema path evaluation."""
-
-    pass
-
-
-def _check_path_function(jobs, path_spec, path_function):
-    """Validate that path function specifies a 1-1 mapping.
-
-    Parameters
-    ----------
-    jobs : iterable of :class:`~signac.contrib.job.Job`
-        A sequence of jobs (instances of :class:`~signac.contrib.job.Job`).
-    path_spec : str
-        The path string that generated the path_function. Displayed in the error message.
-    path_function : callable
-        A callable path generating function.
-
-    Raises
-    -----
-    RuntimeError
-        If paths generated with given path function are not unique.
-
-    """
-    # quick check first
-    links = {path_function(job) for job in jobs}
-    if len(links) != len(jobs):
-        # report all mismatches
-        links = set()
-        for job in jobs:
-            job_path = path_function(job)
-            if job_path in links:
-                logger.debug(f"Generated path '{job_path}' is not unique.")
-            else:
-                links.add(job_path)
-        raise RuntimeError(
-            f"The path specification {path_spec} would result in duplicate links. "
-            "See the debug log for the list. The easiest way to fix "
-            "this is to append the job id to the path "
-            f"specification like '{os.path.join('id', '{job.id}')}'."
-        )
-
-
-def _make_path_function(jobs, path):
-    """Generate a path function for jobs or use ``path`` if it is callable.
-
-    Parameters
-    ----------
-    jobs : iterable of :class:`~signac.contrib.job.Job`
-        A sequence of jobs (instances of :class:`~signac.contrib.job.Job`).
-    path : callable
-        A callable path generating function.
-
-    Returns
-    -------
-    callable
-        Path function for given job and path.
-
-    Raises
-    ------
-    ValueError
-        The path argument must either be ``None``, ``False``, or of type ``str``.
-
-    """
-    if path is None:
-        # Generate a path function based on the schema detected for jobs.
-        path_function = _make_schema_based_path_function(jobs=jobs)
-
-    elif path is False:
-
-        # Just use the job id as path.
-        def path_function(job):
-            """Use job id to construct path.
-
-            Parameters
-            ----------
-            job : :class:`~signac.contrib.job.Job`
-                An instance of :class:`~signac.contrib.job.Job`.
-
-            Returns
-            -------
-            str
-                Job id.
-
-            """
-            return str(job.id)
-
-    elif isinstance(path, str):
-        # Detect keys that are already provided as part of the path specifier
-        # and should therefore be excluded from the 'auto'-part.
-        exclude_keys = [x[1] for x in Formatter().parse(path)]
-
-        # Generate the paths based on the provided selection, excluding any keys
-        # that have already been provided by the user.
-        paths = _make_schema_based_path_function(jobs=jobs, exclude_keys=exclude_keys)
-
-        def path_function(job):
-            """Format a path based on a given string and schema.
-
-            Parameters
-            ----------
-            job : :class:`~signac.contrib.job.Job`
-                An instance of :class:`~signac.contrib.job.Job`.
-
-            Returns
-            -------
-            str
-                Formatted path.
-
-            """
-            try:
-                try:
-                    ret = path.format(job=job, **job.statepoint)
-                except TypeError as error:
-                    if (
-                        str(error)
-                        == "format() got multiple values for keyword argument 'job'"
-                    ):
-                        try:
-                            ret = path.format(job=job)
-                        except KeyError:
-                            raise _SchemaPathEvaluationError(
-                                "You must use fully qualified fields for this path, because the "
-                                "state point contains a key called 'job', e.g.: '{job.sp.job}' "
-                                "instead of '{job}'."
-                            )
-                    ret = path.format(job=job)
-                return _AutoPathFormatter(paths).format(ret, auto=job)
-            except AttributeError as error:
-                raise _SchemaPathEvaluationError(error)
-            except KeyError as error:
-                raise _SchemaPathEvaluationError(f"Unknown key: {error}")
-            except Exception as error:
-                raise _SchemaPathEvaluationError(error)
-
-        # Check that the user-specified path generates a 1-1 mapping
-        _check_path_function(jobs, path_spec=path, path_function=path_function)
-
-    else:
-        raise ValueError(
-            "The path argument must either be `None`, `False`, or of type `str`."
-        )
-
-    return path_function
 
 
 def _check_directory_structure_validity(paths):
